requirements.txt
Django~=4.1.5
djangorestframework~=3.14.0
pymongo==3.12.3
djongo==1.3.6
dnspython==2.3.0
django-cors-headers==3.13.0